Added falling sphere - testing bullet. Fixed 2 warnings - signed unsigned comparison.

+++ camera.h
@@ -40,7 +40,7 @@ public:
 		ComputeMatrices();
 		CameraMatrix = Projection * View;
 
-		for (int i = 0; i < Shader.Name.size(); i++) {
+		for (size_t i = 0; i < Shader.Name.size(); i++) {
 			Shader.Use(Shader.Name[i]);
 			CameraMatrixID = glGetUniformLocation(Shader.ProgramID, "CameraMatrix");
 			glUniformMatrix4fv(CameraMatrixID, 1, GL_FALSE, &CameraMatrix[0][0]);
